refactor(category): move category to entities
